requirements.txt
Flask==2.2.5
azure-storage-blob==12.19.0
cachetools==5.3.2
PyYAML==6.0.1
gunicorn==21.2.0
pytest==7.4.3